add data source repository

+++ flaskd3/externals/mysql/data_source_repository.py
@@ -3,46 +3,82 @@ from flaskd3.externals.mysql import orm
 from datetime import datetime
 from sqlalchemy.exc import IntegrityError, InvalidRequestError, ObjectNotExecutableError
 
-DataSource_Required = ['user_id']
 
-
+# FIXME: this class has anti-pattern, it will be refactored by repository pattern
 class DataSourceRepository(domains.DataSourceRepositoryIF):
 
     def __init__(self, session):
-        # super()
         self.session = session  # set session to connect database
 
     @classmethod
-    def _record_to_model(cls, record):
-        # judge whether dictionary includes the necessary parameters or not
-        def _includes(dic, requires):
-            for require in requires:
-                if require not in dic.keys():
-                    return False
-            return True
-        # guard
-        if not _includes(record,
-                         ['user_id',
-                          'name',
-                          'data_type',
-                          'region',
-                          'id',
-                          'created_at',
-                          'updated_at',
-                          'deleted_at']
-                         ):
-            raise Exception("Fail to convert record to model: datasource")
-        return domains.DataSource(
+    def _bucket_record_to_model(cls, record):
+        return domains.Bucket(
+            record.user_id,
+            record.name,
+            record.region,
+            record.id,
+            record.created_at,
+            record.updated_at,
+            record.deleted_at
+        )
+
+    def _find_bucket_by_id(self, bucket_id):
+        try:
+            record = self.session \
+                .query(orm.Bucket) \
+                .filter_by(id=bucket_id) \
+                .filter(orm.Bucket.deleted_at > datetime.now()) \
+                .first()
+            if record is None:
+                return None
+            return DataSourceRepository._bucket_record_to_model(record)
+        except Exception:
+            return domains.NotFoundError(param='Bucket ID', raw=bucket_id)
+
+    def _object_record_to_model(self, record):
+        # find bucket
+        bucket = self._find_bucket_by_id(record.bucket_id)
+        return domains.DataObject(
+            record.user_id,
+            record.name,
+            record.region,
+            record.id,
+            record.version,
+            bucket,
+            record.created_at,
+            record.updated_at,
+            record.deleted_at
+        )
+
+    def _find_object_by_id(self, object_id):
+        try:
+            record = self.session \
+                .query(orm.Object) \
+                .filter_by(id=object_id) \
+                .filter(orm.Bucket.deleted_at > datetime.now()) \
+                .first()
+            if record is None:
+                return None
+            return self._object_record_to_model(record)
+        except Exception:
+            return domains.NotFoundError(param='Object ID', raw=object_id)
+
+    def _datasource_record_to_model(self, record):
+        ob = self._find_object_by_id(record.object_id)
+        # TODO: I want to set object directly
+        ds = domains.DataSource(
             record.user_id,
             record.name,
             record.data_type,
-            record.region,
+            ob.region,
             record.id,
             None,
             record.created_at,
             record.updated_at,
             record.deleted_at
         )
+        ds.object = ob
+        return ds
 
     # find data source by id from database. This method should return data source object or None
     def find_by_id(self, data_id):
@@ -54,9 +90,9 @@ class DataSourceRepository(domains.DataSourceRepositoryIF):
                 .first()
             if record is None:
                 return None
-            return DataSourceRepository._record_to_model(record)
+            return self._datasource_record_to_model(record)
         except Exception:
-            raise domains.NotFoundError(param=data_id)
+            return domains.NotFoundError(param=data_id)
 
     # find data source by name from database. This method should return data source object or None
     def find_by_name(self, name):
@@ -68,43 +104,40 @@ class DataSourceRepository(domains.DataSourceRepositoryIF):
                 .first()
             if record is None:
                 return None
-            return DataSourceRepository._record_to_model(record)
+            return self._datasource_record_to_model(record)
         except Exception:
-            raise domains.NotFoundError(param=name)
+            return domains.NotFoundError(param=name)
 
     # find data sources by user id from database. This method should return data sources object or None
     def find_by_user_id(self, user_id):
         try:
-            record = self.session \
+            records = self.session \
                 .query(orm.Datum) \
                 .filter_by(user_id=user_id) \
                 .filter(orm.Datum.deleted_at > datetime.now())
-            if record is None:
-                return None
-            # TODO: 複数返ってくるはず
-            return DataSourceRepository._record_to_model(record)
+            if not records:
+                return []
+            record_list = [self._datasource_record_to_model(record) for record in records]
+            return record_list
         except Exception:
-            raise domains.NotFoundError(param=user_id)
+            return domains.NotFoundError(param=user_id)
 
     # save data source with the object to database. This method should return saved data source object or error
     # https://docs.sqlalchemy.org/en/latest/orm/session_basics.html
     def save(self, data_source):
         # NOTE: Call ObjectRepository.save method before call this method
-        def _data_source_to_records(ds):
-            data_source_copied = ds.copy()
-            records = orm.Datum(
-                user_id=data_source_copied.user_id,
-                object_id=data_source_copied.object.id,
-                name=data_source_copied.name,
-                data_type=data_source_copied.data_type.name.lower(),
-                deleted_at=data_source_copied.deleted_at
+        def _data_source_to_record(ds):
+            return orm.Datum(
+                user_id=ds.user_id,
+                name=ds.name,
+                data_type=ds.data_type.name.lower(),
+                object_id=ds.object.id,
+                deleted_at=ds.deleted_at
             )
-            return records
         try:
-            records = _data_source_to_records(data_source)
-            print('records:', records)
+            record = _data_source_to_record(data_source)
             # save record
-            self.session.add(records)
+            self.session.add(record)
             # reflect database
             self.session.commit()
             # TODO: update id, created_at, updated_at
@@ -112,56 +145,74 @@ class DataSourceRepository(domains.DataSourceRepositoryIF):
                 .filter_by(name=data_source.name, user_id=data_source.user_id) \
                 .first()
             self.session.flush()
-            return DataSourceRepository._record_to_model(saved)
+            return self._datasource_record_to_model(saved)
         except IntegrityError as e:
             if 'Duplicate entry' in str(e):
-                raise Exception('Invalid parameter duplicate error occurred: DataSource')
-            raise e
+                return Exception('Invalid parameter duplicate error occurred: DataSource')
+            return e
 
     # delete data source with the object from database. This method should return deleted data source object or error
     # https://docs.sqlalchemy.org/en/latest/core/dml.html
     def delete(self, data_source):
         # NOTE: Call ObjectRepository.delete method after call this method
         def _data_source_to_record(ds):
-            data_source_dict = ds.copy().__dict__
-            return orm.Datum(data_source_dict)
+            return orm.Datum(
+                id=ds.id,
+                user_id=ds.user_id,
+                name=ds.name,
+                data_type=ds.data_type.name.lower(),
+                object_id=ds.object.id,
+                created_at=ds.created_at,
+                updated_at=ds.updated_at
+            )
         try:
             record = _data_source_to_record(data_source)
             # find target data
             found_data = self.session.query(orm.Datum)\
                 .filter_by(id=record.id)\
                 .first()
-            # TODO: handle found_data is None
+            if found_data is None:
+                self.session.flush()
+                return Exception(f'Cannot delete data source because it was not found: {data_source.name}')
             self.session.delete(found_data)
             # reflect database(commit should be called when update, insert, delete)
             self.session.commit()
-            deleted = self.session.query(orm.Datum) \
-                .filter_by(id=found_data.id) \
-                .first()
             self.session.flush()
-            return DataSourceRepository._record_to_model(deleted)
+            found_data.deleted_at = datetime.now()
+            return self._datasource_record_to_model(found_data)
         except (InvalidRequestError, ObjectNotExecutableError) as e:
-            raise Exception(f'Cannot delete {data_source.name} datasource: {str(e)}')
+            return Exception(f'Cannot delete {data_source.name} datasource: {str(e)}')
 
     # update data source with the object to database. This method should return updated data source object or error
     def update(self, data_source):
         def _data_source_to_record(ds):
-            data_source_dict = ds.copy().__dict__
-            del data_source_dict['updated_at']
-            return orm.Datum(data_source_dict)
+            return orm.Datum(
+                id=ds.id,
+                user_id=ds.user_id,
+                name=ds.name,
+                data_type=ds.data_type.name.lower(),
+                object_id=ds.object.id,
+                created_at=ds.created_at,
+                updated_at=ds.updated_at,
+                deleted_at=ds.deleted_at,
+            )
         try:
             record = _data_source_to_record(data_source)
             # find target data
             found_data = self.session.query(orm.Datum) \
                 .filter_by(id=record.id) \
                 .first()
-            # TODO: handle found_data is None
+            if found_data is None:
+                self.session.flush()
+                return Exception(f'Cannot update data source because it was not found: {data_source.name}')
             self.session.merge(record)
             self.session.commit()
             updated = self.session.query(orm.Datum) \
                 .filter_by(id=found_data.id) \
                 .first()
             self.session.flush()
-            return DataSourceRepository._record_to_model(updated)
-        except (InvalidRequestError, ObjectNotExecutableError) as e:
-            raise Exception(f'Cannot update {data_source.name} datasource: {str(e)}')
+            return self._datasource_record_to_model(updated)
+        except IntegrityError as e:
+            if 'Duplicate entry' in str(e):
+                return Exception('Invalid parameter duplicate error occurred: DataSource')
+            return e
